chore: Update NoteModel and EditNoteView to allow editing note title and subtitle
The code changes in NoteModel and EditNoteView modify the title and subtitle properties to be mutable. This allows users to edit the title and subtitle of a note when viewing it in the EditNoteView. This improves the functionality of the app by enabling users to make changes to the note's title and subtitle.